Complete rework of splitpane bar which is now content bar and maintain 1 to 1 ration with contentvuer.
