+++ src/components/viewers/Flatmap.vue
@@ -49,6 +49,7 @@ export default {
     flatmapReadyCall: function (flatmap) {
       let provClone = {id: this.entry.id, prov: this.getFlatmapImp().provenance}; //create clone of provenance and add id
       EventBus.$emit("mapImpProv", provClone); // send clone to context card
+      this.$emit("flatmap-provenance-ready", provClone);
       this.getAvailableTerms();
       if (this.entry.resource === "FunctionalConnectivity"){
         this.flatmapReadyForMarkerUpdates(flatmap);
