+++ src/components/viewers/MultiFlatmap.vue
@@ -237,6 +237,7 @@ export default {
         this.flatmapMarkerZoomUpdate(true, flatmapImp);
         let provClone = {id: this.entry.id, prov: newMapImp.provenance}; //create clone of provenance
         EventBus.$emit("mapImpProv", provClone); // send provenance close to the context card for display
+        this.$emit("flatmap-provenance-ready", provClone);
       }
     },
     getFlatmapImp: function () {
